+++ src/error.rs
@@ -6,7 +6,7 @@
 // ------------------------------------------------------------------------
 use pyo3::{
     create_exception,
-    exceptions::{PyException, PyValueError},
+    exceptions::{PyConnectionError, PyException, PyValueError},
     PyErr,
 };
 
@@ -16,7 +16,7 @@ pub type HttpResult<T> = Result<T, GufoHttpError>;
 pub enum GufoHttpError {
     Request(String),
     Redirect,
-    Connect,
+    Connect(String),
     ValueError(String),
 }
 
@@ -31,14 +31,12 @@ create_exception!(_fast, RequestError, HttpError, "Request error");
 
 create_exception!(_fast, RedirectError, HttpError, "Redirects limit exceeded");
 
-create_exception!(_fast, ConnectError, HttpError, "Connect error");
-
 impl From<GufoHttpError> for PyErr {
     fn from(value: GufoHttpError) -> Self {
         match value {
             GufoHttpError::Request(x) => RequestError::new_err(x),
             GufoHttpError::Redirect => RedirectError::new_err("redirects limit exceeded"),
-            GufoHttpError::Connect => ConnectError::new_err("connect error"),
+            GufoHttpError::Connect(x) => PyConnectionError::new_err(x),
             GufoHttpError::ValueError(x) => PyValueError::new_err(x),
         }
     }
@@ -47,7 +45,7 @@ impl From<GufoHttpError> for PyErr {
 impl From<reqwest::Error> for GufoHttpError {
     fn from(value: reqwest::Error) -> Self {
         if value.is_connect() {
-            return GufoHttpError::Connect;
+            return GufoHttpError::Connect(value.to_string());
         }
         if value.is_redirect() {
             return GufoHttpError::Redirect;